Style: Add a responsive style for HeaderNormalNav
On this commit,
1.   Add a responsive stlye for HeaderNormalNav
1.1. Modify the default style to nested in media query of max-width
1.2. Add a responsive style for min-width of 768px
1.3. Abstract style from Button styled comp. in order to reuse

Now we have a list of button ui for nav over min-width of 768px.

Changes of plan,
since there's no much of responsive styles needed,
I decided to add addtional features, such as DownloadFile,
which supports downloading of Resume and Portfolio, etc..

In result, the next commit will come late,
Resume and Portfolio should be updated before the feature addition.

The next project's coming, better late than never.
It's a very simple to-do app, but with good UX and functionality.
It'd be simultaneously developed with this very project.